server/requirements.txt
Django
pyzmq
redis